refactor : 디렉토리 수정 및 응답 값 수정
