+++ src/main/java/TubeSlice/tubeSlice/domain/user/User.java
@@ -24,6 +24,8 @@ public class User {
 
     private String name;
 
+    private String nickname;
+
     private String email;
 
     private String loginId;
