--- src/main/java/TubeSlice/tubeSlice/domain/user/dto/response/LoginResponseDto.java
@@ -1,14 +1,0 @@
-package TubeSlice.tubeSlice.domain.user.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class LoginResponseDto {
-
-    private Long userId;
-    private String access_token;
-}
